pyproject.toml
[tool.poetry]
name = "manifesto-qa"
version = "0.1.0"
description = ""
authors = ["ed-longbottom <ed.m.longbottom@bbc.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.30.3"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
weaviate-client = "^4.6.3"
requests = "^2.32.2"
pandas = "^2.2.2"
langchain = "^0.2.1"
langchain-weaviate = "^0.0.2"
langchain-community = "^0.2.1"
langchain-openai = "^0.1.8"
streamlit = "^1.36.0"
lark = "^1.1.9"
text-generation = "^0.7.0"
transformers = "^4.44.0"
google-search-results = "^2.4.2"
numexpr = "^2.10.1"
langchainhub = "^0.1.21"
sentencepiece = "^0.2.0"
jinja2 = "^3.1.4"
langchain-huggingface = "^0.0.3"
langchain-mistralai = "^0.1.12"
langchain-cohere = "^0.2.2"
mistralai = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
